rename rssbriefing package
