+++ application.py
@@ -3,7 +3,7 @@
     Module with WSGI callable
 
 """
-from rssbriefing_package import create_app
+from rssbriefing import create_app
 
 application = create_app()
