+++ rssbriefing/auth.py
@@ -4,9 +4,9 @@ import functools
 from flask import Blueprint, current_app as app, flash, g, redirect, render_template, request, session, url_for
 from werkzeug.security import check_password_hash, generate_password_hash
 
-from rssbriefing_package import db
-from rssbriefing_package.db_utils import get_user_by_username
-from rssbriefing_package.models import Users
+from rssbriefing import db
+from rssbriefing.db_utils import get_user_by_username
+from rssbriefing.models import Users
 
 bp = Blueprint('auth', __name__, url_prefix='/auth')
