+++ rssbriefing/briefing_model/preparation.py
