+++ rssbriefing/db_utils.py
@@ -1,4 +1,4 @@
-from rssbriefing_package.models import Users, Feed
+from rssbriefing.models import Users, Feed
 
 
 def get_user_by_id(user_id):
