+++ rssbriefing/scripts/update_all_feeds.py
@@ -1,8 +1,8 @@
 import socket
 
-from rssbriefing_package import create_app
-from rssbriefing_package.feed import get_latest_feed_dict, update_feed_db
-from rssbriefing_package.models import Feed
+from rssbriefing import create_app
+from rssbriefing.feed import get_latest_feed_dict, update_feed_db
+from rssbriefing.models import Feed
 
 # Set timeout in seconds for new socket objects, needed for feedparser connections
 socket.setdefaulttimeout(10)
