+++ tests/test_briefing.py
@@ -1,4 +1,4 @@
-from rssbriefing_package.db_utils import get_feedlist_for_dropdown
+from rssbriefing.db_utils import get_feedlist_for_dropdown
 
 
 def add_one_feed_to_users_feedlist(client):
