+++ tests/test_factory.py
@@ -1,4 +1,4 @@
-from rssbriefing_package import create_app
+from rssbriefing import create_app
 
 
 def test_config():
